client/requirements.txt
Flask==2.3.2
requests==2.31.0
pyjwt==2.10.1